+++ services/session/server/avsession_item.cpp
@@ -155,7 +155,7 @@ int32_t AVSessionItem::DestroyTask()
     }
 
 #ifdef CASTPLUS_CAST_ENGINE_ENABLE
-    SLOGI("Session destroy with castHandle: %{public}lld", castHandle_);
+    SLOGI("Session destroy with castHandle: %{public}lld", (long long)castHandle_);
     if (descriptor_.sessionTag_ != "RemoteCast" && castHandle_ > 0) {
         if (!collaborationNeedNetworkId_.empty()) {
             CollaborationManager::GetInstance().PublishServiceState(collaborationNeedNetworkId_.c_str(),
@@ -854,7 +854,7 @@ int32_t AVSessionItem::RegisterListenerStreamToCast(const std::map<std::string, 
     int64_t castHandle = AVRouter::GetInstance().StartCast(outputDeviceInfo, castServiceNameMapState_, GetSessionId());
     castHandle_ = castHandle;
     castHandleDeviceId_ = deviceInfo.deviceId_;
-    SLOGI("RegisterListenerStreamToCast check handle set to %{public}lld", castHandle_);
+    SLOGI("RegisterListenerStreamToCast check handle set to %{public}lld", (long long)castHandle_);
     CHECK_AND_RETURN_RET_LOG(castHandle != AVSESSION_ERROR, AVSESSION_ERROR, "StartCast failed");
     AVRouter::GetInstance().RegisterCallback(castHandle, cssListener_, GetSessionId());
     AVRouter::GetInstance().SetServiceAllConnectState(castHandle, deviceInfo);
@@ -1022,7 +1022,7 @@ void AVSessionItem::HandleCastValidCommandChange(std::vector<int32_t> &cmds)
 {
     std::lock_guard lockGuard(castControllersLock_);
     SLOGI("send command change event to controller, controller size: %{public}d, cmds size is: %{public}d,",
-        static_cast<int>(castControllers_.size()), cmds.size());
+        static_cast<int>(castControllers_.size()), static_cast<int>(cmds.size()));
     for (auto controller : castControllers_) {
         if (controller != nullptr) {
             controller->HandleCastValidCommandChange(cmds);
@@ -1087,7 +1087,7 @@ int32_t AVSessionItem::StartCast(const OutputDeviceInfo& outputDeviceInfo)
     // unregister pre castSession callback to avoid previous session timeout disconnect influence current session
     if (castHandle_ > 0) {
         if (castHandleDeviceId_ == outputDeviceInfo.deviceInfos_[0].deviceId_) {
-            SLOGI("repeat startcast %{public}lld", castHandle_);
+            SLOGI("repeat startcast %{public}lld", (long long)castHandle_);
             return AVSESSION_ERROR;
         } else {
             SLOGI("cast check with pre cast alive %{public}lld, unregister callback", castHandle_);
@@ -1111,7 +1111,7 @@ int32_t AVSessionItem::SubStartCast(const OutputDeviceInfo& outputDeviceInfo)
     CHECK_AND_RETURN_RET_LOG(castHandle != AVSESSION_ERROR, AVSESSION_ERROR, "StartCast failed");
 
     castHandle_ = castHandle;
-    SLOGI("start cast check handle set to %{public}lld", castHandle_);
+    SLOGI("start cast check handle set to %{public}lld", (long long)castHandle_);
     int32_t ret = AddDevice(static_cast<int32_t>(castHandle), outputDeviceInfo);
     if (ret == AVSESSION_SUCCESS) {
         castHandleDeviceId_ = outputDeviceInfo.deviceInfos_[0].deviceId_;
@@ -1133,7 +1133,7 @@ int32_t AVSessionItem::AddDevice(const int64_t castHandle, const OutputDeviceInf
 
 void AVSessionItem::DealDisconnect(DeviceInfo deviceInfo, bool isNeedRemove)
 {
-    SLOGI("Is remotecast, received disconnect event for castHandle_: %{public}lld", castHandle_);
+    SLOGI("Is remotecast, received disconnect event for castHandle_: %{public}lld", (long long)castHandle_);
     if (isNeedRemove) {
         AVRouter::GetInstance().UnRegisterCallback(castHandle_, cssListener_, GetSessionId());
         AVRouter::GetInstance().StopCastSession(castHandle_);
@@ -1354,7 +1354,7 @@ int32_t AVSessionItem::StopCast()
             AVSessionRadar::GetInstance().StopCastBegin(descriptor_.outputDeviceInfo_, info);
             int64_t ret = AVRouter::GetInstance().StopCast(castHandle_);
             AVSessionRadar::GetInstance().StopCastEnd(descriptor_.outputDeviceInfo_, info);
-            SLOGI("StopCast with unchange castHandle is %{public}lld", castHandle_);
+            SLOGI("StopCast with unchange castHandle is %{public}lld", (long long)castHandle_);
             CHECK_AND_RETURN_RET_LOG(ret != AVSESSION_ERROR, AVSESSION_ERROR, "StopCast failed");
         }
     }
@@ -1386,7 +1386,7 @@ void AVSessionItem::UnRegisterDeviceStateCallback()
 
 void AVSessionItem::StopCastSession()
 {
-    SLOGI("Stop cast session process with castHandle: %{public}lld", castHandle_);
+    SLOGI("Stop cast session process with castHandle: %{public}lld", (long long)castHandle_);
     int64_t ret = AVRouter::GetInstance().StopCastSession(castHandle_);
     DoContinuousTaskUnregister();
     if (ret != AVSESSION_ERROR) {
@@ -1466,7 +1466,7 @@ int32_t AVSessionItem::GetAllCastDisplays(std::vector<CastDisplayInfo>& castDisp
         CHECK_AND_RETURN_RET_LOG(display != nullptr, AVSESSION_ERROR, "display is nullptr");
         auto displayInfo = display->GetDisplayInfo();
         SLOGI("GetAllCastDisplays name: %{public}s, id: %{public}llu",
-            displayInfo->GetName().c_str(), static_cast<uint64_t>(displayInfo->GetDisplayId()));
+            displayInfo->GetName().c_str(), (unsigned long long)displayInfo->GetDisplayId());
         auto flag = Rosen::DisplayManagerLite::GetInstance().GetVirtualScreenFlag(displayInfo->GetDisplayId());
         if (flag == Rosen::VirtualScreenFlag::CAST) {
             SLOGI("ReportCastDisplay start in");
